# perf: parallelize font loading and sync wait it