--- a/package.json
+++ b/package.json
@@ -22,6 +22,8 @@
     "docs:rs": "cargo doc --workspace --document-private-items --no-deps",
     "lint": "eslint editors/vscode/src",
     "lint-fix": "eslint editors/vscode/src --fix",
+    "benches": "cargo bench --workspace",
+    "bench": "cargo bench --workspace --bench",
     "test:grammar": "cd syntaxes/textmate && yarn run test",
     "build:typlite": "cargo build --bin typlite",
     "typlite": "target/debug/typlite",